+++ packages/cli/src/commands/dev/index.ts
@@ -1,0 +1,1 @@
+export {dev as default} from './dev';