+++ packages/cli/src/commands/init/init.ts
@@ -13,7 +13,7 @@ export enum Template {
  * Create a new `@shopify/hydrogen` app.
  */
 export async function init(env: Env) {
-  const {ui, fs, workspace, ...passThroughEnv} = env;
+  const {ui, fs, workspace, hooks, ...passThroughEnv} = env;
 
   const name = await ui.ask('What do you want to name this app?', {
     validate: validateProjectName,
@@ -46,7 +46,7 @@ export async function init(env: Env) {
 
   if (template === Template.None) {
     const context = {name, ...passThroughEnv.context};
-    await app({...passThroughEnv, ui, fs, workspace, context});
+    await app({...passThroughEnv, hooks, ui, fs, workspace, context});
   }
 
   if (template === Template.Default) {
@@ -67,7 +67,11 @@ export async function init(env: Env) {
   }
 
   console.log();
-  workspace.commit().then(() => finish({ui, workspace}));
+
+  hooks.onUpdateFile = async (filePath) => {
+    await ui.printFile(filePath);
+  };
+  hooks.onCommit = finish;
 }
 
 async function finish({ui, workspace}: Pick<Env, 'ui' | 'workspace'>) {