+++ packages/hydrogen/src/hooks/index.ts
@@ -5,3 +5,4 @@ export {useQuery} from '../foundation/useQuery';
 export {useMoney} from './useMoney';
 export {useMeasurement} from './useMeasurement';
 export {useParsedMetafields} from './useParsedMetafields';
+export {useLoadScript} from './useLoadScript';