+++ packages/hydrogen/src/utilities/index.ts
@@ -12,7 +12,7 @@ export {
   addParametersToEmbeddedVideoUrl,
   useEmbeddedVideoUrl,
 } from './video_parameters';
-export {loadScript} from './script_loader';
+export {loadScript} from './load_script';
 export {wrapPromise} from './suspense';
 export {flattenConnection} from './flattenConnection';
 export {isClient} from './isClient';