+++ packages/cli/CHANGELOG.md
@@ -7,6 +7,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+- Adds `dev` command for running Hydrogen apps locally with vite
 - Adds `preview` command for running Hydrogen apps in a local Worker runtime
 - Prevent CLI commands from unintentionally mutating the app code
 - Adds cache support for `preview` command